models: Turn Moodle completion status into float

+++ models/learning_env_class.py
@@ -32,7 +32,7 @@ class MoodleClass(LearningEnvClass):
     """
     Abstract dataclass for a class hosted on Moodle.
     """
-    completion_status: int
+    completion_status: float
 
 
 class LearningEnvClassEnc(JSONEncoder):
